Split manage-prototype routes and handlers
